fix by today statistic and other fixes

+++ analyze.sh
@@ -13,6 +13,8 @@ FILENAME='fail2ban.html'
 WHOISSERVER='http://whois.ens.my/'
 OUTPUTFILE=$OUTPUTLOCATION$FILENAME
 LOG=TRUE
+DATETODAY=$( date )
+DATETODAYDMY=$( date +%d-%m-%Y )
 
 ###HEAD###
 cat > $OUTPUTFILE <<END
@@ -26,34 +28,41 @@ cat > $OUTPUTFILE <<END
 <div id="body">
 <div id="title">
 <h1>Fail2ban Log Analyzer</h1>
+<p>Last Update: $DATETODAY </p>
 </div>
 END
 ###/HEAD###
 
 ###BODY###
-echo '<div id="row"><h2>Banned Count by IP < today ></h2><table>' >> $OUTPUTFILE
+echo "<div id=\"row\"><h2>Banned Count by IP < today $DATETODAYDMY ></h2><table>" >> $OUTPUTFILE
 #grep "Ban " /var/log/fail2ban.log | grep `date +%Y-%m-%d` | awk '{print $NF}' | sort | uniq -c | sort -n > TEMPVAR
 grep "Ban " /var/log/fail2ban.log | grep `date +%Y-%m-%d` | awk -F[\ \:] '{print $10,$8}' | sort | uniq -c | sort -n > TEMPVAR
 I=0
-cat TEMPVAR | while IFS= read -r line
-do
-	if [ $I -eq 0 ]
-	then
-		echo "<tr>" >> $OUTPUTFILE
-	fi
-	IP=$( printf '%s\n' "$line" | awk '{print $2}' )
-	IPCOUNT=$( printf '%s\n' "$line" | awk '{print $1}' )
-	IPTYPE=$( printf '%s\n' "$line" | awk '{print $3}' )
-	echo "<td>"$IPCOUNT"&nbsp;<a href=\"$WHOISSERVER$IP\">"$IP"</a>&nbsp;"$IPTYPE"<td>" >> $OUTPUTFILE
-	if [ $I -eq 5 ]
-	then
-		echo  "</tr>" >> $OUTPUTFILE
-		I=0
-	elif [ $I -lt 5 ]
-	then
-		I=$(( I + 1 ))
-	fi
-done
+##cat TEMPVAR | while IFS= read -r line
+if [ ! -s TEMPVAR ]
+then
+	echo  "<tr><td>No IP banned for this date.</td></tr>" >> $OUTPUTFILE
+else
+	cat TEMPVAR | while read line
+	do
+		if [ $I -eq 0 ]
+		then
+			echo "<tr>" >> $OUTPUTFILE
+		fi
+		IP=$( printf '%s\n' "$line" | awk '{print $2}' )
+		IPCOUNT=$( printf '%s\n' "$line" | awk '{print $1}' )
+		IPTYPE=$( printf '%s\n' "$line" | awk '{print $3}' )
+		echo "<td>"$IPCOUNT"&nbsp;<a href=\"$WHOISSERVER$IP\" target=\"_blank\">"$IP"</a>&nbsp;"$IPTYPE"<td>" >> $OUTPUTFILE
+		if [ $I -eq 5 ]
+		then
+			echo  "</tr>" >> $OUTPUTFILE
+			I=0
+		elif [ $I -lt 5 ]
+		then
+			I=$(( I + 1 ))
+		fi
+	done
+fi
 echo '</table></div>' >> $OUTPUTFILE
 echo '<div id="row"><h2>Banned Count Group by Date and section</h2><table>' >> $OUTPUTFILE
 zgrep -h "Ban " /var/log/fail2ban.log* | awk '{print $5,$1}' | sort | uniq -c | sort -n > TEMPVAR
@@ -88,7 +97,7 @@ do
 	IP=$( printf '%s\n' "$line" | awk '{print $2}' )
 	IPCOUNT=$( printf '%s\n' "$line" | awk '{print $1}' )
 	IPTYPE=$( printf '%s\n' "$line" | awk '{print $3}' )
-	echo "<td>"$IPCOUNT"&nbsp;<a href=\"$WHOISSERVER$IP\">"$IP"</a>&nbsp;"$IPTYPE"<td>" >> $OUTPUTFILE
+	echo "<td>"$IPCOUNT"&nbsp;<a href=\"$WHOISSERVER$IP\" target=\"_blank\">"$IP"</a>&nbsp;"$IPTYPE"<td>" >> $OUTPUTFILE
 	if [ $I -eq 5 ]
 	then
 		echo  "</tr>" >> $OUTPUTFILE
